requirements.txt
numpy
scipy
pandas
Pillow
nltk
scikit-learn
nibabel
nilearn
matplotlib
seaborn
openpyxl
